+++ RELEASE_NOTES.md
@@ -7,5 +7,6 @@
 ## New Features
 
 - Adds a resampler for a stream of samples to resample to a given interval.
+- Adds python bindings for the resampler.
 
 ## Bug Fixes